Move the C shim into a separate file
